+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 
 - Allows `stdin` data to be provided when running commands
 - Adds `create_executable` to which writes a file and adds execute permission
+- Adds `run_interactive` to allow pseudo-terminal interaction
 
 ## v0.2.0